doc: add Rust documentation
