+++ crates/rethnet_evm/src/lib.rs
@@ -1,3 +1,9 @@
+//! The Rethnet EVM
+//!
+//! The Rethnet EVM exposes APIs for running and interacting with a multi-threaded Ethereum
+//! Virtual Machine (or EVM).
+#![warn(missing_docs)]
+
 pub use bytes::Bytes;
 pub use db::layered_db::{LayeredDatabase, RethnetLayer};
 pub use debug::{DatabaseDebug, HasDatabaseDebug};
@@ -9,6 +15,7 @@ pub use revm::{
     ExecutionResult, Log, Return, SpecId, TransactOut, TransactTo, TxEnv, EVM,
 };
 
+/// State mapping of addresses to accounts.
 pub type State = HashMap<H160, Account>;
 
 mod db;
